requirements.txt
django==4.2.10
stripe==2.56.0
django-environ==0.11.2
django-allauth==0.54.0
django-widgets-improved==1.5.0
django-crispy-forms==1.14.0
Pillow==10.3.0
